requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==4.1.7
django-filter==23.1
django-redis==5.2.0
djangorestframework==3.14.0
pytz==2023.3
redis==4.5.4
sqlparse==0.4.3
tzdata==2023.3
